requirements.txt
langchain
langchain-google-genai
python-dotenv
streamlit
unidecode
pydantic
pymupdf
google-generativeai
